rename containers dir to views
