--- src/app/(Home)/domainz/page.tsx
@@ -1,3 +1,0 @@
-import DomainPage from '@/containers/domains-page/domains-page';
-
-export default DomainPage;
