--- src/components/test-component/test-component.tsx
@@ -1,3 +1,0 @@
-export default function TestComponent() {
-  return <div> Testat </div>;
-}
