+++ src/views/domains-page/__fixtures__/domains.ts
