+++ src/views/domains-page/config/domains-page-filters.config.ts
