+++ src/views/domains-page/config/domains-page-query-params.config.ts
