+++ src/views/domains-page/config/domains-table-columns.config.ts
