+++ src/views/domains-page/domains-page-title-badge/domains-page-title-badge.types.ts
