+++ src/views/domains-page/domains-page.tsx
@@ -1,10 +1,10 @@
 import React, { JSXElementConstructor, Suspense } from 'react';
-import DomainPageHeader from '@/containers/domains-page/domains-page-filters/domains-page-filters';
-import DomainsTable from '@/containers/domains-page/domains-table/domains-table';
-import DomainPageTitle from '@/containers/domains-page/domains-page-title/domains-page-title';
-import DomainsPageTitleBadge from '@/containers/domains-page/domains-page-title-badge/domains-page-title-badge';
+import DomainPageHeader from '@/views/domains-page/domains-page-filters/domains-page-filters';
+import DomainsTable from '@/views/domains-page/domains-table/domains-table';
+import DomainPageTitle from '@/views/domains-page/domains-page-title/domains-page-title';
+import DomainsPageTitleBadge from '@/views/domains-page/domains-page-title-badge/domains-page-title-badge';
 import SectionLoadingIndicator from '@/components/section-loading-indicator/section-loading-indicator';
-import { getCachedAllDomains } from '@/containers/domains-page/helpers/get-all-domains';
+import { getCachedAllDomains } from '@/views/domains-page/helpers/get-all-domains';
 import AsyncPropsLoader from '@/components/async-props-loader/async-props-loader';
 
 export default async function DomainsPage() {
