+++ src/views/domains-page/domains-page.types.ts
