+++ src/views/domains-page/domains-table-cluster-cell/__tests__/domains-table-cluster-cell.test.tsx
