+++ src/views/domains-page/domains-table-cluster-cell/domains-table-cluster-cell.tsx
@@ -1,8 +1,8 @@
 'use client';
 import React from 'react';
 
-import type { DomainData } from '@/containers/domains-page/domains-page.types';
-import TableLink from '@/containers/domains-page/domains-table-link/domains-table-link';
+import type { DomainData } from '@/views/domains-page/domains-page.types';
+import TableLink from '@/views/domains-page/domains-table-link/domains-table-link';
 
 import useStyletronClasses from '@/hooks/use-styletron-classes';
 import { cssStyles } from './domains-table-cluster-cell.styles';
