+++ src/views/domains-page/domains-table-domain-name-cell/domains-table-domain-name-cell.styles.ts
