+++ src/views/domains-page/domains-table-domain-name-cell/domains-table-domain-name-cell.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import Image from 'next/image';
 import cadenceIcon from '@/assets/cadence-logo.svg';
 
-import type { DomainData } from '@/containers/domains-page/domains-page.types';
-import TableLink from '@/containers/domains-page/domains-table-link/domains-table-link';
+import type { DomainData } from '@/views/domains-page/domains-page.types';
+import TableLink from '@/views/domains-page/domains-table-link/domains-table-link';
 
 import useStyletronClasses from '@/hooks/use-styletron-classes';
 import { cssStyles } from './domains-table-domain-name-cell.styles';
