+++ src/views/domains-page/domains-table-end-message/domains-table-end-message.type.ts
