+++ src/views/domains-page/domains-table-link/domains-table-link.tsx
