+++ src/views/domains-page/domains-table/domains-table.styles.ts
