+++ src/views/domains-page/domains-table/domains-table.tsx
@@ -6,7 +6,7 @@ import Table from '@/components/table/table';
 import usePageQueryParams from '@/hooks/use-page-query-params/use-page-query-params';
 import sortBy, { SortByReturnValue, toggleSortOrder } from '@/utils/sort-by';
 import useStyletronClasses from '@/hooks/use-styletron-classes';
-import DomainTableEndMessage from '@/containers/domains-page/domains-table-end-message/domains-table-end-message';
+import DomainTableEndMessage from '@/views/domains-page/domains-table-end-message/domains-table-end-message';
 import domainPageQueryParamsConfig from '../config/domains-page-query-params.config';
 
 import { domainTableColumns } from '../config/domains-table-columns.config';
