+++ src/views/domains-page/domains-table/domains-table.types.ts
