+++ src/views/domains-page/helpers/clear-domain-page-filters.ts
