+++ src/views/domains-page/helpers/get-all-domains.ts
